Move About page to its own controller as is
Move the About page to its own controller.
This allow us to split not-shared behaviours
