+++ src/Phansible/Controller/DefaultController.php
@@ -3,9 +3,6 @@
 namespace Phansible\Controller;
 
 use Flint\Controller\Controller;
-use Github\Client;
-use Github\HttpClient\CachedHttpClient;
-use Github\HttpClient\Message\ResponseMediator;
 use Michelf\Markdown;
 use DateTimeZone;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -15,11 +12,6 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  */
 class DefaultController extends Controller
 {
-    /**
-     * @var Client
-     */
-    private $githubClient;
-
     /**
      * @return string
      */
@@ -59,35 +51,5 @@ class DefaultController extends Controller
         return $this->render('docs.html.twig', [
             'content' => $content,
         ]);
-    }
-
-    public function getGithubClient()
-    {
-        if (null === $this->githubClient) {
-            $this->githubClient = new Client(
-                new CachedHttpClient(
-                    ['cache_dir' => __DIR__ . '/../../../app/cache/github-api-cache']
-                )
-            );
-        }
-
-        return $this->githubClient;
-    }
-
-    public function setGithubClient(Client $client)
-    {
-        $this->githubClient = $client;
-        return $this;
-    }
-
-    public function aboutAction()
-    {
-        $response = $this->getGithubClient()
-            ->getHttpClient()
-            ->get('repos/Phansible/phansible/stats/contributors');
-
-        $contributors = ResponseMediator::getContent($response);
-
-        return $this->render('about.html.twig', ['contributors' => $contributors]);
     }
 }
